store/wizard: extract langpack utilities into system module
